Remove unnecessary delegation in comment handling
